feat: 퀘스트 제안 UX 개선 및 쿨다운 안정화
대화 중 퀘스트 요청을 감지해 가능한 경우 즉시 제안 UI를 열고, 불가능할 때는 명확한 안내를 반환해 상호작용 흐름을 개선했습니다. 퀘스트 쿨다운 저장 기준을 UTC 시간으로 변경해 재실행 이후 제안 판정이 꼬이던 문제를 방지했습니다.

Made-with: Cursor

+++ Assets/Scripts/Common/RelationshipSaveManager.cs
@@ -34,7 +34,7 @@ public class RelationshipSaveManager : MonoBehaviour
     private class QuestTimeEntry
     {
         public string questId;
-        public float time;
+        public double time;
     }
 
     [System.Serializable]
@@ -129,7 +129,7 @@ public class RelationshipSaveManager : MonoBehaviour
 
     #region Quest
 
-    public void SaveQuest(string npcId, string questId, Dictionary<string, float> lastQuestTimes)
+    public void SaveQuest(string npcId, string questId, Dictionary<string, double> lastQuestTimes)
     {
         if (string.IsNullOrEmpty(npcId)) return;
         EnsureEntry(npcId);
@@ -155,7 +155,7 @@ public class RelationshipSaveManager : MonoBehaviour
         if (string.IsNullOrEmpty(npcId) || !saveData.TryGetValue(npcId, out var data))
             return null;
 
-        var times = new Dictionary<string, float>();
+        var times = new Dictionary<string, double>();
         if (data.lastQuestTimes != null)
         {
             foreach (var e in data.lastQuestTimes)
@@ -172,7 +172,7 @@ public class RelationshipSaveManager : MonoBehaviour
     public class QuestLoadResult
     {
         public string activeQuestId;
-        public Dictionary<string, float> lastQuestTimes;
+        public Dictionary<string, double> lastQuestTimes;
     }
 
     #endregion
